+++ libs/core-ui/src/index.ts
@@ -107,8 +107,8 @@ export * from "./lib/Highchart/Null";
 export * from "./lib/Highchart/bubbleChart/buildScatterTemplate";
 export * from "./lib/Highchart/bubbleChart/calculateBubbleData";
 export * from "./lib/Highchart/bubbleChart/getBubbleChartOptions";
-export * from "./lib/Highchart/bubbleChart/getCounterfactualsScatter";
-export * from "./lib/Highchart/bubbleChart/getCounterfactualsScatterOption";
+export * from "./lib/Highchart/bubbleChart/getScatterPlot";
+export * from "./lib/Highchart/bubbleChart/getScatterOption";
 export * from "./lib/Highchart/bubbleChart/ChartUtils";
 export * from "./lib/Highchart/IBoxChartState";
 export * from "./lib/Interfaces/IHighchartBubbleData";
